+++ src/lib/Card.svelte
@@ -4,7 +4,7 @@ let classProp: string | undefined = undefined;
 export { classProp as class };
 $: needHeader = !!title || $$slots.header;
 </script>
-<div class:divide-y={needHeader} class="relative divide-gray-200 dark:divide-gray-600 rounded-lg bg-white shadow-xl dark:bg-gray-900 dark:shadow-purple-400/20 ring-1 ring-gray-900/5 {classProp}">
+<div class:divide-y={needHeader} class="relative divide-gray-200 dark:divide-gray-600 rounded-lg bg-white shadow-xl dark:bg-gray-900 dark:shadow-purple-400/20 ring-1 flex flex-col ring-gray-900/5 {classProp}">
   {#if needHeader}
   <div class="px-4 py-5 sm:px-6">
     <div class="flex flex-wrap items-center justify-between sm:flex-nowrap">
@@ -14,7 +14,7 @@ $: needHeader = !!title || $$slots.header;
     </div>
   </div>
   {/if}
-  <div class="px-4 py-5 sm:p-6">
+  <div class="px-4 py-5 sm:p-6 flex flex-col flex-1">
     <slot></slot>
   </div>
 </div>