+++ src/lib/DialogBase.svelte
@@ -3,6 +3,7 @@
   import { blur, scale } from 'svelte/transition';
   import { portal } from './util/portal';
 
+  export let maxWidth = 'max-w-md';
   export let title = '';
   type T = $$Generic;
 
@@ -41,7 +42,7 @@
 
     <div class="fixed inset-0 overflow-y-auto">
       <div class="flex min-h-full items-center justify-center p-4 text-center">
-        <div class="w-full max-w-md transform rounded-2xl bg-white dark:bg-gray-900 p-6 text-left align-middle shadow-xl transition-all" use:dialog.modal transition:scale={{start: 0.5}}>
+        <div class="w-full {maxWidth} transform rounded-2xl bg-white dark:bg-gray-900 p-6 text-left align-middle shadow-xl transition-all" use:dialog.modal transition:scale={{start: 0.5}}>
           <slot name="pretitle"></slot>
           {#if !!title}<h3 class="text-lg font-medium leading-6 text-center">{title}</h3>{/if}
           <slot {close} {open}></slot>