+++ src/lib/data/calling-manager.ts
@@ -1,21 +1,17 @@
 import callings from '$data/callings.json';
 import type { Calling } from '../types';
-import { writable } from 'svelte/store';
 import { id } from '$lib/rolling/id';
+import { renderSafe } from '$lib/md/render';
+import { JsonDataManager } from './base-manager';
 
 const callingList = callings.map(c => ({
   ...c,
-  abilities: c.abilities.map(a => ({...a, id: id()})),
+  abilities: c.abilities.map(a => ({...a, id: id(), desc: renderSafe(a.desc)})),
   equipment: c.equipment.map(e => ({...e, id: id()})),
 })) as Calling[];
 
-class CallingManager {
-  public list = writable(callingList);
-
-  loadList() {
-    // Why are you so empty?
-  }
+class CallingManager extends JsonDataManager<Calling> {
 
 }
 
-export const callingManager = new CallingManager();
+export const callingManager = new CallingManager(callingList);