+++ src/routes/character/[id]/+page.ts
@@ -1,4 +1,0 @@
-
-export async function load({ params }) { 
-  return { id: params.id };
-}