+++ src/routes/character/new/abilities/+page.server.ts
@@ -1,0 +1,9 @@
+import { callingManager } from "$lib/data/calling-manager";
+const callingList = callingManager.getAll();
+import type { PageServerLoad } from './$types';
+
+export const load: PageServerLoad = async function() {
+  return {
+    list: callingList
+  };
+}